refactor: rename api to services folder, move i18n to services
